+++ dscapture.h
@@ -8,9 +8,5 @@ bool capture_grabFrame(uint16_t *frameBuf);
 
 enum {
     FRAMEWIDTH   =256,
-    FRAMEHEIGHT  =192,
-    FRAMESIZE    =FRAMEWIDTH*FRAMEHEIGHT*3,
-    FRAMEBUFSIZE =FRAMESIZE + 0x200,    //extra space for USB maxPacketSize
+    FRAMEHEIGHT  =192
 };
-
-void toRGBA(uint8_t *out, uint16_t *in);
